refactor: rename aletr-type.ts to alert-type.ts
