+++ src/app/health-page/health-panel/health-component/health-component.component.ts
@@ -3,7 +3,7 @@ import { Component, Input } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
 
 import { HealthComponent } from '../../models/health-component.model';
-import { AlertType } from '../../../shared/alert/aletr-type';
+import { AlertType } from '../../../shared/alert/alert-type';
 
 /**
  * A component to render a "health component" object.
