+++ src/app/shared/alert/alert-type.ts
