+++ src/app/shared/alert/alert.component.spec.ts
@@ -8,7 +8,7 @@ import { TranslateModule } from '@ngx-translate/core';
 
 import { AlertComponent } from './alert.component';
 import { createTestComponent } from '../testing/utils.test';
-import { AlertType } from './aletr-type';
+import { AlertType } from './alert-type';
 
 describe('AlertComponent test suite', () => {
