+++ src/app/submission/sections/sherpa-policies/publisher-policy/publisher-policy.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input } from '@angular/core';
 
 import { Policy } from '../../../../core/submission/models/sherpa-policies-details.model';
-import { AlertType } from '../../../../shared/alert/aletr-type';
+import { AlertType } from '../../../../shared/alert/alert-type';
 
 /**
  * This component represents a section that contains the publisher policy informations.
